rename transform --> compose
