+++ compose.js
